56 refactor tests prior to next effort
